+++ boot/ipl3_prod.ld
@@ -36,8 +36,6 @@ SECTIONS {
         KEEP(*(.text.ique_trampoline))
     }
     .text.stage23 0xB0001040  : AT ( 0xA4001040 ) {
-        KEEP(build/loader.o(.text.stage3))
-        . = ALIGN(8);
         __stage2_start = .;
         LONG(__stage2_end - __stage2_start);
         LONG(0);
@@ -46,5 +44,6 @@ SECTIONS {
         KEEP(build/loader.o(.data .data.*))
         . = ALIGN(8);
         __stage2_end = .;
+        KEEP(build/cleanup.o(.text .text.* .data .rodata .rodata.*))
     } > rom
 }
